rfgo-vsto-fastapi/pyproject.toml
[project]
name = "rfgo-vsto-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
    "PyJWT>=2.10.1",
]
